Homework7
